Add decision to standardize metropolitan french
Previously the primary culture of France began spreading to all
available provinces in the French homeland (that were not accepted by a
human player) immediately upon gaining the printing press institution.
This was somewhat at odds with what actually happened historically,
where the 'patois' didn't really start to decline as spoken languages
until the French Revolution and official state imposition of French as
the sole spoken language of the French people.

This changes the French and 'Latin' assimilation mechanics to only work
when the tag has metropolitaine culture, which can only be obtained with
a new decision to annihilate the patois.